add purchase feature

+++ src/main/java/com/be/java/foxbase/db/entity/Book.java
@@ -21,6 +21,6 @@ public class Book {
     String contentUrl;
     String imageUrl;
     String genre;
-    Double price;
+    Long price;
     Double averageRating;
 }
